Entidade Endereco criada com JPA

+++ src/main/java/com/stefanini/dao/PerfilDao.java
@@ -3,6 +3,6 @@ package com.stefanini.dao;
 import com.stefanini.dao.abstracao.GenericDao;
 import com.stefanini.model.Perfil;
 
-public class PerfilDao extends GenericDao<Perfil, Long> {
+public class PerfilDao extends GenericDao<Perfil, Integer> {
 	
 }
